fix long bug

+++ fed_baselines/server_fednova_plus.py
@@ -55,7 +55,7 @@ class FedNovaPlusServer(FedServer):
                 output = self.model(data)
                 loss = loss_func(
                     output,
-                    torch.zeros_like(target) + i,
+                    (torch.zeros_like(target) + i).long(),
                 )
                 self.model.zero_grad()
                 loss.backward()
